User balance to FloatField (from Integer)

+++ client/models.py
@@ -5,7 +5,7 @@ from django.dispatch import receiver
 
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    balance = models.IntegerField(default=50)
+    balance = models.FloatField(default=50)
     oil = models.IntegerField(default=0)
     avatar = models.ImageField(upload_to="avatars", blank=True)
     stat_tower = models.IntegerField(default=0)
